Refactor remotes/fixup.go large functions
