+++ remotes/fixup.go
@@ -20,10 +20,13 @@ import (
 func FixupBundle(ctx context.Context, b *bundle.Bundle, ref reference.Named, resolver remotes.Resolver, opts ...FixupOption) error {
 	logger := log.G(ctx)
 	logger.Debugf("Fixing up bundle %s", ref)
+
+	// Configure the fixup and the even loop
 	cfg, err := newFixupConfig(b, ref, resolver, opts...)
 	if err != nil {
 		return err
 	}
+
 	events := make(chan FixupEvent)
 	eventLoopDone := make(chan struct{})
 	defer func() {
@@ -38,85 +41,61 @@ func FixupBundle(ctx context.Context, b *bundle.Bundle, ref reference.Named, res
 		}
 	}()
 
+	// Fixup invocation images
 	if len(b.InvocationImages) != 1 {
 		return fmt.Errorf("only one invocation image supported for bundle %q", ref)
 	}
 	if b.InvocationImages[0].BaseImage, err = fixupImage(ctx, b.InvocationImages[0].BaseImage, cfg, events, cfg.invocationImagePlatformFilter); err != nil {
 		return err
 	}
+	// Fixup images
 	for name, original := range b.Images {
 		if original.BaseImage, err = fixupImage(ctx, original.BaseImage, cfg, events, cfg.componentImagePlatformFilter); err != nil {
 			return err
 		}
 		b.Images[name] = original
 	}
+
 	logger.Debug("Bundle fixed")
 	return nil
 }
 
-func fixupImage(ctx context.Context, baseImage bundle.BaseImage, cfg fixupConfig, events chan<- FixupEvent, platformFilter platforms.Matcher) (_ bundle.BaseImage, retErr error) {
+func fixupImage(ctx context.Context, baseImage bundle.BaseImage, cfg fixupConfig, events chan<- FixupEvent, platformFilter platforms.Matcher) (bundle.BaseImage, error) {
 	log.G(ctx).Debugf("Fixing image %s", baseImage.Image)
 	ctx = withMutedContext(ctx)
-	progress := &progress{}
-	originalSource := baseImage.Image
-	notifyEvent := func(eventType FixupEventType, message string, err error) {
-		events <- FixupEvent{
-			DestinationRef: cfg.targetRef,
-			SourceImage:    originalSource,
-			EventType:      eventType,
-			Message:        message,
-			Error:          err,
-			Progress:       progress.snapshot(),
-		}
-	}
-	defer func() {
-		if retErr != nil {
-			notifyEvent(FixupEventTypeCopyImageEnd, "", retErr)
-		}
-	}()
+	notifyEvent, progress := makeEventNotifier(events, baseImage.Image, cfg.targetRef)
+
 	notifyEvent(FixupEventTypeCopyImageStart, "", nil)
+	// Fixup Base image
 	fixupInfo, err := fixupBaseImage(ctx, &baseImage, cfg.targetRef, cfg.resolver)
 	if err != nil {
-		return bundle.BaseImage{}, err
+		return notifyError(notifyEvent, err)
 	}
-
 	if fixupInfo.sourceRef.Name() == fixupInfo.targetRepo.Name() {
 		notifyEvent(FixupEventTypeCopyImageEnd, "Nothing to do: image reference is already present in repository"+fixupInfo.targetRepo.String(), nil)
 		return baseImage, nil
 	}
-	sourceRepoOnly, err := reference.ParseNormalizedNamed(fixupInfo.sourceRef.Name())
+
+	sourceFetcher, err := makeSourceFetcher(ctx, cfg.resolver, fixupInfo.sourceRef.Name())
 	if err != nil {
-		return bundle.BaseImage{}, err
-	}
-	f, err := cfg.resolver.Fetcher(ctx, sourceRepoOnly.Name())
-	if err != nil {
-		return bundle.BaseImage{}, err
-	}
-	sourceFetcher := newSourceFetcherWithLocalData(f)
-	if err := fixupPlatforms(ctx, &baseImage, &fixupInfo, sourceFetcher, platformFilter); err != nil {
-		return bundle.BaseImage{}, err
+		return notifyError(notifyEvent, err)
 	}
 
-	// Prepare the copier
-	copier, err := newDescriptorCopier(ctx, cfg.resolver, sourceFetcher, fixupInfo.targetRepo.String(), notifyEvent, fixupInfo.sourceRef)
+	// Fixup platforms
+	if err := fixupPlatforms(ctx, &baseImage, &fixupInfo, sourceFetcher, platformFilter); err != nil {
+		return notifyError(notifyEvent, err)
+	}
+
+	// Prepare and run the copier
+	walkerDep, cleaner, err := makeManifestWalker(ctx, sourceFetcher, notifyEvent, cfg, fixupInfo, progress)
 	if err != nil {
-		return bundle.BaseImage{}, err
+		return notifyError(notifyEvent, err)
 	}
-	descriptorContentHandler := &descriptorContentHandler{
-		descriptorCopier: copier,
-		targetRepo:       fixupInfo.targetRepo.String(),
-	}
-	ctx, cancel := context.WithCancel(ctx)
-	scheduler := newErrgroupScheduler(ctx, cfg.maxConcurrentJobs, cfg.jobsBufferLength)
-	defer func() {
-		cancel()
-		scheduler.drain()
-	}()
-	walker := newManifestWalker(notifyEvent, scheduler, progress, descriptorContentHandler)
-	walkerDep := walker.walk(scheduler.ctx(), fixupInfo.resolvedDescriptor, nil)
+	defer cleaner()
 	if err = walkerDep.wait(); err != nil {
-		return bundle.BaseImage{}, err
+		return notifyError(notifyEvent, err)
 	}
+
 	notifyEvent(FixupEventTypeCopyImageEnd, "", nil)
 	return baseImage, nil
 }
@@ -173,10 +152,10 @@ func fixupBaseImage(ctx context.Context,
 	baseImage *bundle.BaseImage,
 	targetRef reference.Named, //nolint: interfacer
 	resolver remotes.Resolver) (imageFixupInfo, error) {
-	err := checkBaseImage(baseImage)
-	if err != nil {
-		err := fmt.Errorf("invalid image %q: %s", baseImage.Image, err)
-		return imageFixupInfo{}, err
+
+	// Check image references
+	if err := checkBaseImage(baseImage); err != nil {
+		return imageFixupInfo{}, fmt.Errorf("invalid image %q: %s", baseImage.Image, err)
 	}
 	targetRepoOnly, err := reference.ParseNormalizedNamed(targetRef.Name())
 	if err != nil {
@@ -184,14 +163,14 @@ func fixupBaseImage(ctx context.Context,
 	}
 	sourceImageRef, err := reference.ParseNormalizedNamed(baseImage.Image)
 	if err != nil {
-		err = fmt.Errorf("%q is not a valid image reference for %q", baseImage.Image, targetRef)
-		return imageFixupInfo{}, err
+		return imageFixupInfo{}, fmt.Errorf("%q is not a valid image reference for %q: %s", baseImage.Image, targetRef, err)
 	}
 	sourceImageRef = reference.TagNameOnly(sourceImageRef)
+
+	// Try to fetch the image descriptor
 	_, descriptor, err := resolver.Resolve(ctx, sourceImageRef.String())
 	if err != nil {
-		err = fmt.Errorf("failed to resolve %q, push the image to the registry before pushing the bundle: %s", sourceImageRef, err)
-		return imageFixupInfo{}, err
+		return imageFixupInfo{}, fmt.Errorf("failed to resolve %q, push the image to the registry before pushing the bundle: %s", sourceImageRef, err)
 	}
 	digested, err := reference.WithDigest(targetRepoOnly, descriptor.Digest)
 	if err != nil {
@@ -205,27 +184,4 @@ func fixupBaseImage(ctx context.Context,
 		sourceRef:          sourceImageRef,
 		targetRepo:         targetRepoOnly,
 	}, nil
-}
-
-func checkBaseImage(baseImage *bundle.BaseImage) error {
-	switch baseImage.ImageType {
-	case "docker":
-	case "oci":
-	case "":
-		baseImage.ImageType = "oci"
-	default:
-		return fmt.Errorf("image type %q is not supported", baseImage.ImageType)
-	}
-
-	switch baseImage.MediaType {
-	case ocischemav1.MediaTypeImageIndex:
-	case ocischemav1.MediaTypeImageManifest:
-	case images.MediaTypeDockerSchema2Manifest:
-	case images.MediaTypeDockerSchema2ManifestList:
-	case "":
-	default:
-		return fmt.Errorf("image media type %q is not supported", baseImage.ImageType)
-	}
-
-	return nil
 }
